move 'isVerified' to utils
